fix(client): post[id] 리뷰리스트와 아이템 분리
